# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Pets Api",
        "version": "0.0.1"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a User",
                "description": "Creates an User on the Database",
                "operationId": "da8643013c2941f5ed7f8a9c9b4a7f7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm Password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Created"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login With a registered User",
                "operationId": "535ffeda4b78916efd658b2844ea897a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get User Data",
                "description": "Get the authenticated user data",
                "operationId": "57d51f3fa328d534a6ee8562484d0687",
                "responses": {
                    "200": {
                        "description": "Authenticated User data",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Log out an user",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Logged Out",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "description": "Refresh a User Token before that expires",
                "operationId": "6e53f8dae00ca458918da553749b9ae4",
                "responses": {
                    "200": {
                        "description": "Content with a new token will be returned",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pets/my_pets": {
            "get": {
                "tags": [
                    "My Pets"
                ],
                "summary": "Get User Pets",
                "description": "Get the authenticated user pets",
                "operationId": "6377272e7ee2fda416eb79012777e585",
                "responses": {
                    "200": {
                        "description": "Will return all user pets",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pet not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pets": {
            "post": {
                "tags": [
                    "My Pets"
                ],
                "summary": "Register a New Pet",
                "description": "Creates a new pet for authenticated user",
                "operationId": "9523721805529edb47b46332d56f2826",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age",
                                    "breed",
                                    "specie"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Pet Name",
                                        "type": "string"
                                    },
                                    "age": {
                                        "description": "Pet Age",
                                        "type": "string"
                                    },
                                    "breed": {
                                        "description": "Pet Breed",
                                        "type": "string"
                                    },
                                    "specie": {
                                        "description": "Pet Specie (dog, cat, bird...)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pet Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pets/{petId}": {
            "get": {
                "tags": [
                    "My Pets"
                ],
                "summary": "Get User Single Pet",
                "description": "Get the authenticated user single pet profile",
                "operationId": "585e7638b2477aa2613bf6d11c8f9556",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "description": "ID of the pet to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Will return a pet profile",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated or unauthorized when a user try to see an animal that isn't yours",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pet not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "My Pets",
            "description": "My Pets"
        }
    ]
}